+++ src/store/aiInfra/slices/aiModel/selectors.ts
@@ -70,32 +70,38 @@ const modelContextWindowTokens = (id: string, provider: string) => (s: AIProvide
   return model?.contextWindowTokens;
 };
 
-const modelExtendControls = (id: string, provider: string) => (s: AIProviderStoreState) => {
+const modelExtendParams = (id: string, provider: string) => (s: AIProviderStoreState) => {
   const model = getEnabledModelById(id, provider)(s);
 
-  return model?.settings?.extendControls;
+  return model?.settings?.extendParams;
 };
 
-const isModelHasExtendControls = (id: string, provider: string) => (s: AIProviderStoreState) => {
-  const controls = modelExtendControls(id, provider)(s);
+const isModelHasExtendParams = (id: string, provider: string) => (s: AIProviderStoreState) => {
+  const controls = modelExtendParams(id, provider)(s);
 
   return !!controls && controls.length > 0;
 };
 
-const isModelHasBuiltinSearch = (id: string, provider: string) => (s: AIProviderStoreState) => {
+const modelBuiltinSearchImpl = (id: string, provider: string) => (s: AIProviderStoreState) => {
   const model = getEnabledModelById(id, provider)(s);
 
-  return !!model?.settings?.searchImpl;
+  return model?.settings?.searchImpl;
+};
+
+const isModelHasBuiltinSearch = (id: string, provider: string) => (s: AIProviderStoreState) => {
+  const searchImpl = modelBuiltinSearchImpl(id, provider)(s);
+
+  return !!searchImpl;
 };
 
 const isModelHasBuiltinSearchConfig =
   (id: string, provider: string) => (s: AIProviderStoreState) => {
-    const model = getEnabledModelById(id, provider)(s);
+    const searchImpl = modelBuiltinSearchImpl(id, provider)(s);
 
     return (
-      !!model?.settings?.searchImpl &&
+      !!searchImpl &&
       [ModelSearchImplement.Tool, ModelSearchImplement.Params].includes(
-        model?.settings?.searchImpl as ModelSearchImplement,
+        searchImpl as ModelSearchImplement,
       )
     );
   };
@@ -113,12 +119,13 @@ export const aiModelSelectors = {
   isModelHasBuiltinSearch,
   isModelHasBuiltinSearchConfig,
   isModelHasContextWindowToken,
-  isModelHasExtendControls,
+  isModelHasExtendParams,
   isModelLoading,
   isModelSupportReasoning,
   isModelSupportToolUse,
   isModelSupportVision,
+  modelBuiltinSearchImpl,
   modelContextWindowTokens,
-  modelExtendControls,
+  modelExtendParams,
   totalAiProviderModelList,
 };