+++ src/tools/web-browsing/Portal/Footer.tsx
@@ -1,6 +1,6 @@
 import { ActionIcon, Icon } from '@lobehub/ui';
 import { Button } from 'antd';
-import { LucideBotMessageSquare, LucideNotepadText } from 'lucide-react';
+import { LucideNotepadText, PlusSquareIcon } from 'lucide-react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
@@ -8,34 +8,37 @@ import { useChatStore } from '@/store/chat';
 import { chatPortalSelectors, chatSelectors } from '@/store/chat/selectors';
 
 const Footer = () => {
-  const [messageId, isAIGenerating, triggerAIMessage, summaryPluginContent] = useChatStore((s) => [
+  const [messageId, isAIGenerating, triggerAIMessage, saveSearchResult] = useChatStore((s) => [
     chatPortalSelectors.toolMessageId(s),
     chatSelectors.isAIGenerating(s),
     s.triggerAIMessage,
-    s.summaryPluginContent,
+    s.saveSearXNGSearchResult,
   ]);
-  const { t } = useTranslation('portal');
+
+  const { t } = useTranslation('tool');
 
   return (
     <Flexbox gap={8} horizontal paddingBlock={12} paddingInline={12}>
       <Button
-        icon={<Icon icon={LucideBotMessageSquare} />}
-        loading={isAIGenerating}
-        onClick={() => {
-          triggerAIMessage({ parentId: messageId });
-        }}
-      >
-        {t('actions.genAiMessage')}
-      </Button>
-      <ActionIcon
-        icon={LucideNotepadText}
+        icon={<Icon icon={LucideNotepadText} />}
         loading={isAIGenerating}
         onClick={() => {
           if (!messageId) return;
 
-          summaryPluginContent(messageId);
+          triggerAIMessage({});
         }}
-        title={t('actions.summaryTooltip')}
+      >
+        {t('search.summaryTooltip')}
+      </Button>
+      <ActionIcon
+        icon={PlusSquareIcon}
+        loading={isAIGenerating}
+        onClick={() => {
+          if (!messageId) return;
+
+          saveSearchResult(messageId);
+        }}
+        title={t('search.createNewSearch')}
       />
     </Flexbox>
   );